Run traefik in its own systemd unit
When we reload hub config, we do not want to interrupt
user sessions. By separating hub and proxy processes,
restarting the hub process no longer restarts the proxy
process.

However, we need to share credentials between hub and
proxy so they can communicate securely. The script currently
in `tbjh/__init__.py` should be idempotent as much
as possible. So we put creds into a separate JSON file
that isn't regenerated unless we must.